Implement support for exhaust cursor in OP_MSG
This commit allows client applications to specify a cursor to be
exhaustAllowed. When specified, the corresponding exhaustAllowed
flagbit in the OP_MSG will be set on requests to the server. Future
getMore requests will also not be required, as the next batch of
results will already be sitting on the connection.

JAVA-2950